Add unity:prefabKeepObject/Collider
Using prefab imports nukes the SuperObject, SuperCollider, and Collider
metadata imported from unity, and while custom properties are broadcast
onto your prefab, there is no way to retain the object geometry or tile
metadata.

Add two new unity attributes, unity:prefabKeepObject and
unity:prefabKeepCollider, that on import, checks if set on the object,
and if so, creates a copy of the SuperObject or the SuperColliderComponent
and Collider2D and copies them onto the root node in the prefab.