Implement unity:prefabKeepObject
Currently, using prefab imports nukes the SuperObject and
the SuperCustomProperties components.  While custom properties are
broadcast, it may still be desirable to allow access to arbitrary custom
properties by name at runtime.

Add a new unity attributes, unity:prefabKeepObject that on import,
checks if set on the object, and if so, creates a copy of the
SuperObject and SuperCustomProperties onto the root node of the prefab.